fix: extract wallet, remove gasLimit
